long issue?

+++ trl/trainer/grpo_trainer.py
@@ -1089,7 +1089,7 @@ class GRPOTrainer(Trainer):
                 if self.is_fsdp_enabled else nullcontext()
             ):
                 unwrapped_model.to(torch.bfloat16)
-                all_outputs = unwrapped_model.generate_batch(prompt_inputs.input_ids, generation_config=self.generation_config) 
+                all_outputs = unwrapped_model.generate_batch(prompt_inputs.input_ids.long(), generation_config=self.generation_config) 
             completion_ids = [output.generated_tokens for output in all_outputs.values()]
             completion_ids = [torch.tensor(ids, device=device) for ids in completion_ids]
             completion_ids = pad(completion_ids, padding_value=self.processing_class.pad_token_id)
